doc: Updated LICENE and technologies section
The LICENSE stated MIT, even though we have switched to AGPL. This was updated.

The technologies section now features all the latest technologies in use, and
longer and better explainations on what they do in the project. Also included
is the configuration files in use for each tool, and who provides it and how.

The section about gosbtones-scripts have been added to explain how this tool
hides away the configuration of some of the tooling, or how it provides defaults
that should not be manually updated.