build: transform the guideline into a webpage
By transforming the guidelines project into a webpage, now the
contents of the documentation can be hosted through GitHub Pages
as with any other source of documentation.
Also, better navigation and styling can be achieved if needed,
specially with image and others. tyle of the page resembles that
of the generated TypeDoc, making it easy for the developers to
navigate all the documentation pages now.
A page about the guidelines project itself has been added, the
pages for the community had been splitted, and miner styling
has been done through all content pages.

BREAKING CHANGE: The whole project structure has changed.